add optional add_to_scope argument to pyquery()
This allows to insert additional variables into the scope the jello query runs in.

This is helpful for example when using jello as Jinja/Ansible filter and you want
to access other variables like facts or the inventory.

Does not affect running jello on the commandline, only usable when calling the
jello module from other python programs.