Wrap user query python code in function using ast
The user query is compiled and the resulting body is placed within
a function definition.  The last statement, if an expression, is wrapped
in a Return statement.  The resulting function is compiled/exec'ed to
create a function object.  This function is then called.

The user query may now contain return, yield, and yield from statements.
If the user query contains yield or yield from statements the resulting
function is a generator.

If the function returns a iterator or is a generator the results will be
placed in a list before serializing to json.