support flattened, streaming output ndjson: option -F
option -F will flatten the output, formatting and printing each item of the
query's list, iterator, or generator output as a distinct line of
newline-delimited json.  Writing data is streamed so that the output data
set is not held in memory in its entirety.

An error occurs if the returned output is not a list or iterator, or if the
user's query does not yield as a generator.